🛠  MULE-IOS [MODIFIED] ios iosDeviceAction function moved
