+++ src/ios/functions.js
@@ -69,7 +69,23 @@ const iosEmulatorList = (runtimeKey) => {
   });
 };
 
+const iosDeviceAction = (flag, udid) => {
+  exec(
+    `sh ${SCRIPT_PREFIX}options_ios_device.sh -${flag} ${udid}`,
+    (err, stdout, stderr) => {
+      if (err) {
+        console.error(err);
+        exit;
+      } else {
+        if (stdout) {
+        }
+      }
+    }
+  );
+};
+
 module.exports = {
   iosRuntimeList,
   iosEmulatorList,
+  iosDeviceAction,
 };
