+++ src/ios/index.js
@@ -1,6 +1,11 @@
-const { iosRuntimeList, iosEmulatorList } = require("./functions");
+const {
+  iosRuntimeList,
+  iosEmulatorList,
+  iosDeviceAction,
+} = require("./functions");
 
 module.exports = {
   iosRuntimeList,
   iosEmulatorList,
+  iosDeviceAction,
 };
